+++ src/main/java/de/maxhenkel/gravestone/tileentity/GraveStoneTileEntity.java
@@ -13,7 +13,7 @@ import net.minecraft.network.NetworkManager;
 import net.minecraft.network.play.server.SUpdateTileEntityPacket;
 import net.minecraft.tileentity.TileEntity;
 
-public class GraveStoneTileEntity extends TileEntity implements IInventory {
+public class GraveStoneTileEntity extends TileEntity {
 
     public static final int INVENTORY_SIZE = 127;
     private Inventory inventory;
@@ -34,6 +34,10 @@ public class GraveStoneTileEntity extends TileEntity implements IInventory {
         this.deathTime = 0L;
         this.playerUUID = "";
         this.renderHead = true;
+    }
+
+    public Inventory getInventory() {
+        return inventory;
     }
 
     @Override
@@ -89,67 +93,7 @@ public class GraveStoneTileEntity extends TileEntity implements IInventory {
 
     @Override
     public CompoundNBT getUpdateTag() {
-        return this.write(new CompoundNBT());
-    }
-
-    @Override
-    public int getSizeInventory() {
-        return inventory.getSizeInventory();
-    }
-
-    @Override
-    public ItemStack getStackInSlot(int index) {
-        return inventory.getStackInSlot(index);
-    }
-
-    @Override
-    public ItemStack decrStackSize(int index, int count) {
-        return inventory.decrStackSize(index, count);
-    }
-
-    @Override
-    public void setInventorySlotContents(int index, ItemStack stack) {
-        this.inventory.setInventorySlotContents(index, stack);
-    }
-
-    @Override
-    public int getInventoryStackLimit() {
-        return inventory.getInventoryStackLimit();
-    }
-
-    @Override
-    public void openInventory(PlayerEntity player) {
-        this.inventory.openInventory(player);
-    }
-
-    @Override
-    public void closeInventory(PlayerEntity player) {
-        this.inventory.closeInventory(player);
-    }
-
-    @Override
-    public boolean isItemValidForSlot(int index, ItemStack stack) {
-        return inventory.isItemValidForSlot(index, stack);
-    }
-
-    @Override
-    public void clear() {
-        this.inventory.clear();
-    }
-
-    @Override
-    public ItemStack removeStackFromSlot(int index) {
-        return this.inventory.removeStackFromSlot(index);
-    }
-
-    @Override
-    public boolean isEmpty() {
-        return inventory.isEmpty();
-    }
-
-    @Override
-    public boolean isUsableByPlayer(PlayerEntity player) {
-        return inventory.isUsableByPlayer(player);
+        return write(new CompoundNBT());
     }
 
     public String getPlayerName() {