+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed GUIs not showing tooltips
+- Fixed gravestone items can be picked up with a hopper